Move start out to scheduler implementation
